refactor: extract routing config, rename user serialization, and improve profile reactivity
