+++ examples/auth_flow/lib/domain/user.dart
@@ -8,8 +8,8 @@ class User {
   final String name;
   final String email;
 
-  factory User.fromJson(Map<String, dynamic> json) =>
-      User(id: json['id'] as String, name: json['name'] as String, email: json['email'] as String);
+  factory User.fromMap(Map<String, dynamic> map) =>
+      User(id: map['id'] as String, name: map['name'] as String, email: map['email'] as String);
 
-  Map<String, dynamic> toJson() => {'id': id, 'name': name, 'email': email};
+  Map<String, dynamic> toMap() => {'id': id, 'name': name, 'email': email};
 }
