+++ examples/auth_flow/lib/ui/profile_page.dart
@@ -1,17 +1,12 @@
 import 'package:auth_flow/notifiers/auth_notifier.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_solidart/flutter_solidart.dart';
 
 class ProfilePage extends StatelessWidget {
   const ProfilePage({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final user = AuthNotifier.provider.of(context).currentUser.value;
-
-    if (user == null) {
-      return const Scaffold(body: Center(child: Text('No user found')));
-    }
-
     return Scaffold(
       appBar: AppBar(
         backgroundColor: Theme.of(context).colorScheme.inversePrimary,
@@ -27,13 +22,23 @@ class ProfilePage extends StatelessWidget {
       ),
       body: Padding(
         padding: const EdgeInsets.all(12),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text('User ID: ${user.id}'),
-            Text('User Name: ${user.name}'),
-            Text('User Email: ${user.email}'),
-          ],
+        child: SignalBuilder(
+          builder: (_, _) {
+            final user = AuthNotifier.provider.of(context).currentUser.value;
+
+            if (user == null) {
+              return const Center(child: Text('No user found'));
+            }
+
+            return Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text('User ID: ${user.id}'),
+                Text('User Name: ${user.name}'),
+                Text('User Email: ${user.email}'),
+              ],
+            );
+          },
         ),
       ),
     );
